+++ src/shared/parameter_v2/c_parameter/example_config.h
@@ -2,7 +2,7 @@
 #include <memory.h>
 #include <stdlib.h>
 
-#include "shared/parameter_v2/config/examples/parameter.h"
+#include "shared/parameter_v2/c_parameter/parameter.h"
 
 ///////////////////////////////////////////////////////////////////////
 //                 Generated C Config Structs                        //
