+++ src/shared/parameter_v2/c_parameter/parameter.c
@@ -1,4 +1,4 @@
-#include "shared/parameter_v2/config/examples/parameter.h"
+#include "shared/parameter_v2/c_parameter/parameter.h"
 
 #include <memory.h>
 #include <stdlib.h>
